make DictBuilder private
