+++ cpp/src/arrow/python/CMakeLists.txt
@@ -47,7 +47,6 @@ set(ARROW_PYTHON_SRCS
   builtin_convert.cc
   common.cc
   config.cc
-  dict.cc
   helpers.cc
   init.cc
   io.cc
@@ -91,7 +90,6 @@ install(FILES
   builtin_convert.h
   common.h
   config.h
-  dict.h
   helpers.h
   init.h
   io.h
