+++ cpp/src/arrow/python/python_to_arrow.h
@@ -22,7 +22,6 @@
 
 #include "arrow/api.h"
 
-#include "arrow/python/dict.h"
 #include "arrow/python/numpy_interop.h"
 #include "arrow/python/sequence.h"
 
@@ -35,6 +34,41 @@ extern PyObject* pyarrow_deserialize_callback;
 
 namespace arrow {
 namespace py {
+
+/// Constructing dictionaries of key/value pairs. Sequences of
+/// keys and values are built separately using a pair of
+/// SequenceBuilders. The resulting Arrow representation
+/// can be obtained via the Finish method.
+class DictBuilder {
+ public:
+  explicit DictBuilder(MemoryPool* pool = nullptr) : keys_(pool), vals_(pool) {}
+
+  /// Builder for the keys of the dictionary
+  SequenceBuilder& keys() { return keys_; }
+  /// Builder for the values of the dictionary
+  SequenceBuilder& vals() { return vals_; }
+
+  /// Construct an Arrow StructArray representing the dictionary.
+  /// Contains a field "keys" for the keys and "vals" for the values.
+
+  /// \param list_data
+  ///    List containing the data from nested lists in the value
+  ///   list of the dictionary
+  ///
+  /// \param dict_data
+  ///   List containing the data from nested dictionaries in the
+  ///   value list of the dictionary
+  arrow::Status Finish(std::shared_ptr<Array> key_tuple_data,
+                       std::shared_ptr<Array> key_dict_data,
+                       std::shared_ptr<Array> val_list_data,
+                       std::shared_ptr<Array> val_tuple_data,
+                       std::shared_ptr<Array> val_dict_data,
+                       std::shared_ptr<Array>* out);
+
+ private:
+  SequenceBuilder keys_;
+  SequenceBuilder vals_;
+};
 
 arrow::Status SerializeSequences(std::vector<PyObject*> sequences,
                                  int32_t recursion_depth,
